requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.3.0
decorator==4.4.2
EasyProcess==0.2.10
future==0.18.2
gast==0.2.2
google-api-core==1.16.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gym==0.17.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
moviepy==1.0.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
Pillow==7.0.0
proglog==0.1.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pytz==2019.3
PyVirtualDisplay==0.2.5
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.43.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
